Added support for automatic spawning, and hooked some of FishList into Main
Note: raft and lighting currently disabled, 10-unit cube at 0,0,0 to test odd cameara offset

+++ includes/physics.h
@@ -1,2 +1,2 @@
-float deltaTime();
+int deltaTime();
 void updateTime();
